+++ CHANGELOG.md
@@ -10,17 +10,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `build_model` function to create zero-shot models. ([#584](https://github.com/jina-ai/finetuner/pull/584))
+
 - Use latest Hubble with `notebook_login` support. ([#576](https://github.com/jina-ai/finetuner/pull/576))
 
 ### Removed
 
 ### Changed
 
-- Convert package to namespace package. ([#581](https://github.com/jina-ai/finetuner/pull/581))
+- Use the run config model from `finetuner-stubs` to create the run config. ([#579](https://github.com/jina-ai/finetuner/pull/579))
 
 - Use `device` parameter to replace `cpu` to align with docarray. ([#577](https://github.com/jina-ai/finetuner/pull/577))
 
 - Update the open clip model names in the table of the backbones. ([#580](https://github.com/jina-ai/finetuner/pull/580))
+
+- Show progress while encode batches of documents. ([#586](https://github.com/jina-ai/finetuner/pull/586))
+
+- Change `device` as an optional parameter when calling `get_model`. ([#586](https://github.com/jina-ai/finetuner/pull/586))
 
 ### Fixed
 
@@ -31,6 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Embed three tasks as three Google Colab notebooks in documentation. ([#583](https://github.com/jina-ai/finetuner/pull/583))
 
 - Unify documentation related to cloud storage as Jina AI Cloud. ([#582](https://github.com/jina-ai/finetuner/pull/582))
+
+- Replace `hub.jina.ai` with `cloud.jina.ai`. ([#587](https://github.com/jina-ai/finetuner/pull/587))
 
 
 ## [0.6.3] - 2022-10-13